Create environment with ghc, ghcid, and select packages
For now, adding new packages will mean editing shell.nix. That
seems fine to me. It's possible I'll want to define build.nix
at some point and have shell.nix use it.

There seem to be lots of ways to write these things. For example,
this is based on an example from
nixos.wiki/wiki/Development_environment_with_nix-shell, but this
is the first time I've heard of mkShell.