Solved Day 9, Part 1
Finally! Not sure what the problem was; I think
I just perpetually fixed one problem and introduced
another. This commit finally fixed everything at once!

In the end, I think the issue was with how I tried to generalize
access to source operands and the destination operand. They
need to be treated separately.

This commit includes a crude breakpoint feature for eval; I'm
not sure it works. I should either strip it or fix it.